Rename and convert dfnshelper to typescript
